extract db info
